[generator][python] Refactored: moved stages to stages.py
