+++ tools/python/maps_generator/generator/stages.py
@@ -1,12 +1,92 @@
 import logging
 import os
 import shutil
+import subprocess
 
+from . import settings
+from .env import WORLD_NAME, WORLDS_NAMES
 from .gen_tool import run_gen_tool
+from .osmtools import osmconvert, osmupdate
+from ..utils.file import download_file, is_verified
 from ..utils.file import symlink_force
+from ..utils.md5 import write_md5sum, md5
 
 logger = logging.getLogger("maps_generator")
 
+
+def download_planet(planet):
+    download_file(settings.PLANET_URL, planet)
+    download_file(settings.PLANET_MD5_URL, md5(planet))
+
+
+def convert_planet(tool, in_planet, out_planet, output=subprocess.DEVNULL,
+                   error=subprocess.DEVNULL):
+    osmconvert(tool, in_planet, out_planet, output=output, error=error)
+    write_md5sum(out_planet, md5(out_planet))
+
+
+def stage_download_and_convert_planet(env, **kwargs):
+    if not is_verified(settings.PLANET_PBF):
+        download_planet(settings.PLANET_PBF)
+
+    convert_planet(env[settings.OSM_TOOL_CONVERT],
+                   settings.PLANET_PBF, settings.PLANET_O5M,
+                   output=env.get_subprocess_out(),
+                   error=env.get_subprocess_out())
+    os.remove(settings.PLANET_PBF)
+    os.remove(md5(settings.PLANET_PBF))
+
+
+def stage_update_planet(env, **kwargs):
+    tmp = settings.PLANET_O5M + ".tmp"
+    osmupdate(env[settings.OSM_TOOL_UPDATE], settings.PLANET_O5M, tmp,
+              output=env.get_subprocess_out(),
+              error=env.get_subprocess_out(),
+              **kwargs)
+    os.remove(settings.PLANET_O5M)
+    os.rename(tmp, settings.PLANET_O5M)
+    write_md5sum(settings.PLANET_O5M, md5(settings.PLANET_O5M))
+
+
+def stage_preprocess(env, **kwargs):
+    run_gen_tool(env.gen_tool,
+                 out=env.get_subprocess_out(),
+                 err=env.get_subprocess_out(),
+                 intermediate_data_path=env.intermediate_path,
+                 osm_file_type="o5m",
+                 osm_file_name=settings.PLANET_O5M,
+                 node_storage=env.node_storage,
+                 preprocess=True,
+                 **kwargs)
+
+
+def stage_features(env, **kwargs):
+    extra = {}
+    if not env.production:
+        extra["no_ads"] = True
+    if any(x not in WORLDS_NAMES for x in env.countries):
+        extra["split_by_polygons"] = True
+        extra["generate_packed_borders"] = True
+    if any(x == WORLD_NAME for x in env.countries):
+        extra["generate_world"] = True
+
+    run_gen_tool(
+        env.gen_tool,
+        out=env.get_subprocess_out(),
+        err=env.get_subprocess_out(),
+        data_path=env.data_path,
+        intermediate_data_path=env.intermediate_path,
+        osm_file_type="o5m",
+        osm_file_name=settings.PLANET_O5M,
+        node_storage=env.node_storage,
+        user_resource_path=env.user_resource_path,
+        dump_cities_boundaries=True,
+        cities_boundaries_data=env.cities_boundaries_path,
+        generate_features=True,
+        emit_coasts=True,
+        **extra,
+        **kwargs
+    )
 
 def run_gen_tool_with_recovery_country(env, *args, **kwargs):
     if "data_path" not in kwargs or "output" not in kwargs:
